beginnings of permission admin
I changed the table structure (again). It's a choice between two of
efficiency, integrity and no data duplication - I picked the last two.

We pass the permission regexps around as binaries and store them like
that too, just as we do for all other string-y datums.

I was toying with the idea of making set_permissions take a
#permission record instead of two separate regexp params, so that it
is easy to add more permissions later. But that is inconsistent with
the rest of the rabbit_access_control API, which is "flat", and
complicates the code in rabbit_control and the rabbit_tests.

--HG--
branch : bug20173