Done. Testing the latency between the call to conserve_memory and when the msg gets processed shows this is benficial:
In the erlang client, hammer in a few million messages, with no consumer. This causes the channel mailbox to get pretty big. Without the higher priority can demonstrate a delay of over a second before the conserve_message gets processed. With this, it's a mere fraction of that.

--HG--
branch : bug21046