When stopping rabbitmq server, kill any rabbitmq-owned epmd process
epmd is the Erlang Port Mapper Daemon, which maps Erlang node names to
TCP port numbers.  It is shared by all users on a system, and started
automatically by the Erlang start-up script if not already running.
As a result, epmd will often be started when the rabbitmq server
starts, and so will be run as the rabbitmq user.  This process will
linger even after the rabbitmq server is stopped.

This can have surprising consequences: The deluser program in Debian
Sid checks whether any processes owned by the given user are running,
and will fail if so.

This change addresses the problem in this init script.  When stopping
the rabbitmq server, it will also kill any rabbitmq-owned epmd
process.

--HG--
branch : bug20978